requirements.txt
pandas
scikit-learn
maxminddb-geolite2
streamlit
